chore: refactor stats page according to new specs
